feat(pg-compat): split modules and add DID/JWT authentication
Split oversized bridge.rs (1634 lines) and handler.rs (1043 lines) into
focused module directories, and add DID+JWT authentication to the PG wire
protocol server.

bridge/ module: mod.rs (types + dispatcher), dml.rs, ddl.rs, expr.rs,
params.rs, tests.rs

handler/ module: mod.rs (core execution), protocol.rs (simple/extended
query handlers), catalog.rs (system catalog), cascade.rs (cascade delete),
auth.rs (DID+JWT StartupHandler)

Authentication uses username=DID, password=JWT with anonymous fallback.
Identity is threaded through execute_graphql to QueryRequest::with_identity().
